requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
boto3==1.34.0
requests==2.31.0
python-dotenv==1.0.0
markdown==3.5.1
pandas==2.1.4
aiofiles==23.2.0
httpx==0.25.2
pydantic==2.5.0
python-multipart==0.0.6 